requirements.txt
python_dotenv ~= 0.19.2
requests ~= 2.27.1
requests_toolbelt ~= 0.9.1
spotipy ~= 2.19.0
youtube_search_python ~= 1.5.3
